fix(Firebase): fixes Navigation navigate issue
