+++ packages/client/src/components/Footer/Footer.component.js
@@ -65,7 +65,7 @@ export const Footer = () => {
                     <Link to="/cookies-policies">Cookies policies</Link>
                   </li>
                   <li>
-                    <Link to="/contactsUs">Contact us</Link>
+                    <Link to="/contacts-us">Contact us</Link>
                   </li>
                   <li>
                     <Link to="/FAQs">FAQs</Link>
