+++ packages/client/src/containers/ContactUs.styles.css
@@ -1,5 +1,5 @@
 * {
-  background-color: #e4c1b1;
+  /*background-color: #e4c1b1;*/ /* it sets this background color everywhere, so need to replace*/
   margin: 0 auto;
 }
 .contactUs-wrapper {
